See #10

+++ migrations/20131203022425-create-users-projects-relation.js
@@ -10,5 +10,9 @@ exports.up = function(db, callback) {
 };
 
 exports.down = function(db, callback) {
-  db.dropTable(n, callback);
+  if (process.env.DROP_TABLES == "true") {
+    db.dropTable(n, callback);
+  } else {
+    callback();
+  }
 };
